Move data table utils into separate file
