+++ addon/components/polaris-data-table.js
@@ -7,62 +7,12 @@ import { assign } from '@ember/polyfills';
 import ContextBoundEventListenersMixin from 'ember-lifeline/mixins/dom';
 import ContextBoundTasksMixin from 'ember-lifeline/mixins/run';
 import layout from '../templates/components/polaris-data-table';
+import { measureColumn, getPrevAndCurrentColumns } from '../utils/data-table';
 
 function elementLookup(selector) {
   return computed(function() {
     return this.element.querySelector(selector);
   });
-}
-
-function measureColumn(tableData) {
-  return function(column, index) {
-    let {
-      tableLeftVisibleEdge,
-      tableRightVisibleEdge,
-      firstVisibleColumnIndex,
-      fixedColumnWidth,
-    } = tableData;
-    let width = column.offsetWidth;
-    let leftEdge = column.offsetLeft - fixedColumnWidth;
-    let rightEdge = leftEdge + width;
-    let leftEdgeIsVisible = isEdgeVisible(
-      leftEdge,
-      tableLeftVisibleEdge,
-      tableRightVisibleEdge
-    );
-    let rightEdgeIsVisible = isEdgeVisible(
-      rightEdge,
-      tableLeftVisibleEdge,
-      tableRightVisibleEdge
-    );
-    let isCompletelyVisible =
-      leftEdge < tableLeftVisibleEdge && rightEdge > tableRightVisibleEdge;
-    let isVisible =
-      isCompletelyVisible || leftEdgeIsVisible || rightEdgeIsVisible;
-    if (isVisible) {
-      tableData.firstVisibleColumnIndex = Math.min(
-        firstVisibleColumnIndex,
-        index
-      );
-    }
-    return { leftEdge, rightEdge, isVisible };
-  };
-}
-
-function isEdgeVisible(position, start, end) {
-  let minVisiblePixels = 30;
-  return (
-    position >= start + minVisiblePixels && position <= end - minVisiblePixels
-  );
-}
-
-function getPrevAndCurrentColumns(tableData, columnData) {
-  const { firstVisibleColumnIndex } = tableData;
-  const previousColumnIndex = Math.max(firstVisibleColumnIndex - 1, 0);
-  const previousColumn = columnData[previousColumnIndex];
-  const currentColumn = columnData[firstVisibleColumnIndex];
-
-  return { previousColumn, currentColumn };
 }
 
 /**
